Make history reflect invoice OR quote
Note: There has been a semantic change to the history table
See bamboo_system_files/application/models/invoice_histories_model.php
and bamboo_system_files/application/views/invoices/view.php for details.

This means old databases will have incorrecty history.
If this is a problem, it is a simple change to either this code or the
database to fix it.