Add the paynow via paypal integration.
This adds integration to paypal and google checkout.

I have not tested the google checkout functionality.

I did not write the majority of this patch, but as the forums are
gone I don't know the original author.

If you are the author and want credit I will happily amend the patch
to give you credit. -- Micheas